# Cloud/contracts/framebase-api-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Framebase Development API",
    "version": "1.2.0",
    "description": "Private development cloud contract for Framebase Phase 4 organization and recovery."
  },
  "paths": {
...
    "/v1/blobs/upload-initiate": {
      "post": {
        "summary": "Initiate a signed direct fixture upload",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/blobs/upload-complete": {
      "post": {
        "summary": "Verify an uploaded fixture blob",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/blobs/multipart/initiate": {
      "post": {
        "summary": "Create or resume a private multipart original upload",
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/v1/blobs/multipart/{uploadId}/parts/{partNumber}": {
      "put": {
        "summary": "Upload one bounded authenticated multipart part",
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/v1/blobs/multipart/{uploadId}/complete": {
      "post": {
        "summary": "Assemble a complete multipart original privately in R2",
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/v1/blobs/multipart/{uploadId}/confirm": {
      "post": {
        "summary": "Accept device checksum verification of a completed multipart original",
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/v1/blobs/{id}/verification-download": {
      "get": {
        "summary": "Issue a short-lived verification-only private download",
        "security": [{ "bearerAuth": [] }]
      }
    },
...
    "/v1/blobs/{id}/download": {
      "get": {
        "summary": "Issue a signed direct download",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}